Add add_run command and tests
