+++ tests/test_data/results_provider_test_data.py
@@ -16,16 +16,6 @@ TEST_UPLOAD_RESULTS_FLOW_IDS = [
     "add_results_failed",
     "close_run_failed",
 ]
-TEST_GET_SUITE_ID_PROMPTS_USER_TEST_DATA = [
-    (True, 10, 1, "Adding missing suites to project Fake project name.", False),
-    (True, 10, -1, "Adding missing suites to project Fake project name.", True),
-    (False, -1, -1, FAULT_MAPPING["no_user_agreement"].format(type="suite"), False),
-]
-TEST_GET_SUITE_ID_PROMPTS_USER_IDS = [
-    "user agrees",
-    "user agrees, fail to add suite",
-    "used does not agree",
-]
 TEST_ADD_MISSING_SECTIONS_PROMPTS_USER_TEST_DATA = [
     (
         True,
@@ -87,24 +77,6 @@ TEST_ADD_MISSING_TEST_CASES_PROMPTS_USER_IDS = [
     "user agrees, test cases added",
     "user agrees, test cases not added",
     "used does not agree",
-]
-TEST_GET_SUITE_ID_SINGLE_SUITE_MODE_BASELINES_TEST_DATA = [
-    (([], "Could not get suites"), -1, -1, "Could not get suites"),
-    (([10], ""), -1, 1, ""),
-    (
-        ([10, 11, 12], ""),
-        -1,
-        -1,
-        FAULT_MAPPING["not_unique_suite_id_single_suite_baselines"].format(
-            project_name="Fake project name"
-        ),
-    ),
-]
-
-TEST_GET_SUITE_ID_SINGLE_SUITE_MODE_BASELINES_IDS = [
-    "get_suite_ids fails",
-    "get_suite_ids returns one ID",
-    "get_suite_ids returns more than one ID",
 ]
 
 TEST_REVERT_FUNCTIONS_AND_EXPECTED = [
